Seed plugins via a rake task
When someone installs Intercity, we want to give them access to plugins. Since
we use a database to store them in for now, we also need to seed that DB.

This Rake task will be called upon a intercity-docker bootstrap call (So
everytime someone installs or updates their IC instance), and so it will always
be up to date.

To keep it in sync with dev, and make sure it works correctly, we also use this
task for dev seeding

Signed-off-by: Jeroen van Baarsen <jeroenvanbaarsen@gmail.com>