Implement backups for apps
This adds support for running backups for added integrations. If an integration
supports backups (Postgresql, Redis and Mysql do), we allow the user to create
"backups" of this information.

At this point we do not clean the backups, but we do make sure they are
transfered over to the Intercity server. When we then run the nightly IC
backups, we include the backups created by the apps.

The backups run every hour for now. We can change this in the future.

We also make a difference in manually created backups, and backups that are
created by the system. This way we can later make a distinction between the two
when rotating the backups.

Signed-off-by: Jeroen van Baarsen <jeroenvanbaarsen@gmail.com>