Add automated healthchecks
The healthchecks should run every 30 minutes to check if the server is still
accessible.

If the server is accessible, we also take the time to sync the Dokku version,
since it might be out of sync if the user manually updated their Dokku instance.

When the server is down, we want to give the user the possibility to force a new
healthcheck, you can do so from the Server is Down page.

Fixes: #28

Signed-off-by: Jeroen van Baarsen <jeroenvanbaarsen@gmail.com>