Implemented basic framework for documentation
This will allow us to write extensive documentation in Markdown, and keep it
versioned inside the app. We render it in a pretty way inside the app itself.

The benefit of this, you can also browse the documentation on Github, your
editor and inside the app itself.

This includes code-highlighting.

Signed-off-by: Jeroen van Baarsen <jeroenvanbaarsen@gmail.com>