Remove the service container when we remove an app.
At the moment we keep the linked service containers running. So if you have a
lot of apps added/removed there are a ton of zombie containers.

By making sure we unlink from the app, and then destroy the service we can free
up the needed resources on the server.

Fixes #80

Signed-off-by: Jeroen van Baarsen <jeroenvanbaarsen@gmail.com>