Add fake SSH Executioner for tests
Currently we have to stub out the entire `SshExecution` class, in order not to
actually try to hit a server.

By adding a `FakeSshExecutioner` as a stand-in for the actual lib, we can run
tests without hitting an external server, but still determine the result we want
to get back.

We can improve this later on by also recording what commands have been send to
the `FakeSshExecutioner`.

Signed-off-by: Jeroen van Baarsen <jeroenvanbaarsen@gmail.com>