Redesign onboarding and show errors with fields
**Redesign onboarding**
When we moved to Bulma we did forget to style the onboarding page. I added new
styling to that page

**Show errors with fields**
Before the Bulma upgrade, we used bootstrap and bootstrap-form gem. This was
making sure the fields where in red and with the error message near it. Since
we moved to Bulma, we did not show any feedback to the user when there was an
error.

I added styling for that case downside is, it has to be done with a splash of
JS. This is because of the complex'ish html structure for fields with errors in
Bulma

Fixes #118