Don't schedule backups if server is down (#151)
In the current situation we schedule backups always if the backup service is
enabled. This means that for a server that is down, backups will never finish,
and create a lot of backup entries that are running forever.

By skipping servers that are marked as down, we don't have this issue anymore.