Redesign / Improve env var page
Currently there is a problem where environment variables which are super long
are ruining the layout.

With this fix the env vars are shown in a disabled text field, with a cursor
indicating that it is just read only. When you click the field, it will select
the content within for easy copying.

Fixes: #155