Move settings to setting page
We want to have as little settings in a config file as possible. We still want
to provide certain config option for the end user. In order to accomplish this
we added a Settings page.

In the future when we add new settings, and they are required for the working of
the system, we force users to visit the settings page and fill in the missing
settings.

We probably want to improve that system in the future, but for a first pass its
good enough (tm)