When linking a service to an app, I want to get more feedback
Currently we just redirect the user back to the App overview page, and directly
mark the service as linked. This is not true, since linking a service can take
anywhere between a second up to 30 seconds or more.

In the new situation we show the user that we are linking the service, and poll
to see if the service has been linked.

Fixes: #8