Show the server resources on the server overview page
In order to determine what kind of servers you have running, it helps to see the
resources a server has on the overview page.

We fetch this information when we install the server. For servers that have been
installed before this update, we'll fetch the information in the first
Healthcheck.

Fixes: #199