Initial Enable LetsEncrypt widget on domains page for app (#260)
This commit includes an initial feature for easily enabling, disabling, and refreshing LetsEncrypt on applications. This feature makes use of the Dokku letsencrypt plugin: https://github.com/dokku/dokku-letsencrypt.

There is a BETA UI element on the Domains page of an app where you can:

* enter your LetsEncrypt email (for ACME expire notifications and authentication)
* enable LetsEncrypt
* disable LetsEncrypt
* refresh/update an active LetsEncrypt configuration. for example when domains were added or removed

Under the hood, this feature:

* set the appropriate ENV var on the app with the LetsEncrypt account
* installs the LetsEncrypt plugin (if not installed)
* runs "dokku letsencrypt" on the app on install, and when refresh is used

The actions performed are currently synchronous instead if run in jobs. This way it's easier to give a response to the user and indicate if LetsEncrypt is enabled or disabled. This in absence of a good server state/SSH output logging in Intercity.