Docker setup for production deployment (#265)
This commit finalizes an initial set of changes for easily deploying Intercity to production. The method that we finally ended up with, was the ability to easily wget a script from our Git repository, so that someone can run that on a server and be sure to have everything set up.

This works via the updated Dockerfile, docker-compose.yml, and scripts/bootstrap.sh. The Dockerfile is now updated to build a production-ready image. The new docker-compose.yml file contains all the service definitions to run Intercity, redis, postgres, and letsencrypt security on production. The scripts/bootstrap.sh script can be downloaded from Git onto a clean server and run to do the full initial setup. The script will also interactively ask for any configuration variables on startup.

Changes:
* Update ruby to 2.5.1
* Update Dockerfile to build a production image
* Update docker-compose.yml file with service definitions to run on production
* Read ENV vars from intercity.env file on production servers
* Add scripts/bootstrap.sh to easily install or update Intercity on a server
* Remove that SECRET_KEY_BASE is stored in Redis
* Set default UTF-8 LANG in production Docker container
* Add changelog entries for updated Bundler and docker deployment